fix(test) mv Test_convertLibWithScanner to library_test.go
